Coinertia analysis with identity matrices

Performs a co-inertia of the triplets \((Q_1,K_1,R_0)\) and \((Q_2,K_2,R_0)\) with identity matrices \(K_1, K_2, R_0\).

coinertiaI(X, Y, fast = TRUE)

Arguments

X

Species matrix X.

Y

Species Matrix Y.

fast

If "TRUE" only return the row scores of Y.

Details

Argument fast is used to return only the row scores of Y in function permutest.coca, which speeds the permutation test considerably.

Value

If fast = TRUE, a matrix of row scores for matrix Y (see scores below). If fast = FALSE a list with the following components:

weights

A list with components X and Y containing the left and right singular vectors respectivley of the SVD on the triplets.

scores

A list with components X and Y, containing the row scores of the X and Y species matrices. These are the result of a matrix multiplication of X by the left singular vectors and Y by the right singular vectors.

lambda

the Eigenvalues of the analysis (the square of the singular values from the SVD.

call

the matched function call.

References

Doledec, S and Chessel, D. (1994) Co-inertia analysis: a method for studying species-environment relationships. Freshwater Biology 31, 277–294.

Author

Original Matlab code by C.J.F. ter Braak and A.P. Schaffers. R port by Gavin L. Simpson.

Note

This function is not meant to be called directly by the user. If you wish to use it study the code in permutest.coca to see how it should be called.

See also