Summarizing Co-CA Model Fits

summary methods for classes "predcoca" and "symcoca". These provide a summary of the main results of a Co-Correspondence Analysis model.

# S3 method for predcoca
summary(object, axes = NULL, ...)

# S3 method for symcoca
summary(object, ...)

Arguments

object
an object of class "predcoca" or "symcoca". Generally the result of a call to coca.

axes
numeric; how many axes to summarise? The default is to display 6 axes or all available axes, whichever is the smaller.

...
arguments to be passed to other methods.

Value

A list with the some of the following components:

cocaScores

The site and/or species scores for the axes requested.

call

The call used to fit the model.

lambda

The eigenvalues for the axes requested. Not for predcoca.simpls.

namY, namX

the names of the response and predictor either supplied by the user or derived from the original call.

loadings

a list with two components loadings1 and loadings2, which refer to the response and the predictor matrices respectively. (Only for predictive CoCA models.)

varianceExp

a list with components Yblock and Xblock containing the amount of variance explained on each CoCA axis in the response and the predictor respectively. (Only for predictive CoCA models.)

totalVar

a list with components Yblock and Xblock containing the total variance in the response and the predictor data sets respectively

inertia

a list with components total and residual containing the total and residual inertia (variance) in the response and the predictor matrices of a symmetric CoCA model. (Only for symmetric CoCA models.)

scaling

the scaling used/requested. (Only for symmetric CoCA models.)

Author

Gavin L. Simpson

See also

Examples

od <- options(digits = 4)
## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")
#> 
#> Removed some species that contained no data in: beetles, plants 
summary(bp.sym)
#> 
#> Symmetric Co-Correspondence Analysis
#> 
#> Call: symcoca(y = y, x = x, n.axes = n.axes, R0 = weights, symmetric =
#> symmetric, nam.dat = nam.dat)
#> 
#> Inertia:
#>          Total Explained Residual
#> beetles:  5.70      5.60     0.09
#> plants:   6.16      6.07     0.09
#> 
#> Eigenvalues:
#>   COCA 1    COCA 2    COCA 3    COCA 4    COCA 5    COCA 6    COCA 7    COCA 8  
#> 5.91e-01  3.16e-01  2.03e-01  1.04e-01  6.69e-02  6.11e-02  5.26e-02  4.45e-02  
#>   COCA 9   COCA 10   COCA 11   COCA 12   COCA 13   COCA 14   COCA 15   COCA 16  
#> 3.51e-02  2.31e-02  2.18e-02  1.43e-02  1.34e-02  1.10e-02  1.02e-02  8.39e-03  
#>  COCA 17   COCA 18   COCA 19   COCA 20   COCA 21   COCA 22   COCA 23   COCA 24  
#> 7.44e-03  4.85e-03  4.15e-03  3.45e-03  2.82e-03  2.42e-03  2.26e-03  1.78e-03  
#>  COCA 25   COCA 26   COCA 27   COCA 28   COCA 29  
#> 1.30e-03  1.06e-03  8.53e-04  2.93e-04  5.56e-05  

## Predictive CoCA
bp.pred <- coca(beetles ~ ., data = plants)
#> 
#> Removed some species that contained no data in: beetles, plants 
summary(bp.pred, axes = 1:2)
#> 
#> Predictive Co-Correspondence Analysis
#> 
#> Call: predcoca.simpls(y = y, x = x, R0 = weights, n.axes = n.axes,
#> nam.dat = nam.dat)
#> 
#> Percentage Variance Explained:
#> 
#> Y-block: variance explained in beetles (response) 
#>              Comp 1  Comp 2  Comp 3  Comp 4  Comp 5  Comp 6  Comp 7  Comp 8
#> Individual:   8.18    6.38    5.37    6.91    3.72    5.53    4.18    3.59 
#> Cumulative:   8.18   14.56   19.93   26.84   30.56   36.10   40.27   43.87 
#>              Comp 9  Comp 10  Comp 11  Comp 12  Comp 13  Comp 14  Comp 15
#> Individual:   4.02    3.35     4.50     2.46     3.33     2.56     3.21  
#> Cumulative:  47.89   51.24    55.73    58.19    61.52    64.08    67.29  
#>              Comp 16  Comp 17  Comp 18  Comp 19  Comp 20  Comp 21  Comp 22
#> Individual:   3.32     1.71     2.66     1.96     1.76     1.73     2.17  
#> Cumulative:  70.60    72.31    74.97    76.93    78.69    80.42    82.59  
#>              Comp 23  Comp 24  Comp 25  Comp 26  Comp 27  Comp 28  Comp 29
#> Individual:   1.44     1.81     1.34     2.49     1.56     3.54     2.85  
#> Cumulative:  84.02    85.84    87.18    89.67    91.22    94.76    97.61  
#> 
#> X-block: variance explained in plants (predictor) 
#>              Comp 1    Comp 2    Comp 3    Comp 4    Comp 5    Comp 6  
#> Individual:   21.7114   16.2242   12.8109    5.4318    7.2780    3.7984
#> Cumulative:   21.7114   37.9356   50.7466   56.1784   63.4564   67.2548
#>              Comp 7    Comp 8    Comp 9    Comp 10   Comp 11   Comp 12 
#> Individual:    4.3291    4.2956    2.8320    2.8737    1.7862    2.4498
#> Cumulative:   71.5839   75.8795   78.7115   81.5852   83.3714   85.8212
#>              Comp 13   Comp 14   Comp 15   Comp 16   Comp 17   Comp 18 
#> Individual:    1.5557    1.6480    1.3222    1.1388    1.6991    0.8933
#> Cumulative:   87.3769   89.0249   90.3471   91.4859   93.1850   94.0783
#>              Comp 19   Comp 20   Comp 21   Comp 22   Comp 23   Comp 24 
#> Individual:    0.8779    0.8904    0.8645    0.6373    0.7220    0.5669
#> Cumulative:   94.9562   95.8466   96.7111   97.3484   98.0704   98.6373
#>              Comp 25   Comp 26   Comp 27   Comp 28   Comp 29 
#> Individual:    0.6544    0.3002    0.2774    0.0944    0.0362
#> Cumulative:   99.2918   99.5920   99.8694   99.9638  100.0000
options(od)