Co-inertia analysis

Performs a co-inertia of the triplets \((Q_1,K_1,R_0)\) and \((Q_2,K_2,R_0)\).

fitCoinertia(X, Dp, Y, Dq, Dn, n.axes)

<!-- %\method{print}{coinertia}(x, axes = c(1:min(6, x$n.axes)), -->
<!-- %      digits = max(3, getOption("digits") - 3), \ldots) -->

# S3 method for class 'fitCoinertia'
summary(object, ...)

<!-- %\method{print}{summary.coinertia}(x, digits = max(3, -->
<!-- %      getOption("digits") - 3), \ldots) -->

Arguments

X

\(Q_1\), matrix of expected abundances under row-column independence in the original Y species matrix when treated as a contingency table.

Dp

\(K_1\), species (column) weights for X.

Y

\(Q_2\), matrix of expected abundances under row-column independence in the original X species matrix when treated as a contingency table.

Dq

\(K_2\), species (column) weights for Y.

Dn

site weights \(R_0\).

n.axes

number of axes to calculate the co-inertia analysis for.

object

an object of class coinertia.

axes

the number of axes to display when printing.

...

arguments passed to other functions. Currently ignored.

Value

A list with the following components:

U1

column weights of X.

U2

column weights of Y.

X1

rowscores of X.

X2

rowscores of Y.

lambda

the Eigenvalues (squares of the singular values).

n.axes

number of axes requested.

call

the matched function call.

References

Ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence Analysis: a new ordination method to relate two community compositions. Ecology 85(3), 834–846

Doledec, S and Chessel, D. (1994) Co-inertia analysis: a method for studying species-environment relationships. Freshwater Biology 31, 277–294

Author

Original Matlab code by C.J.F. ter Braak and A.P. Schaffers. R port by Gavin L. Simpson.

Note

This function is not yet meant to be called directly by the user. If you wish to use it directly, see the function definition for symcoca which demonstrates how to prepare the relevant input matrices.

Note that in this function, X corresponds to the input matrix y and Y corresponds to the input matrix x in symcoca. Confusing! This will be changed in a future release but for now the arguments follow those of the original Matlab code - perhaps a little too closely!

See also