Get Species or Site Scores from an Ordination

Function to access either species or site scores for specified axes in co-correspondence analysis ordination methods.

# S3 method for predcoca
scores(x, choices = c(1, 2),
       display = c("sites","species"), ...)

# S3 method for symcoca
scores(x, choices = c(1, 2),
       display = c("sites","species"), scaling = FALSE, ...)

Arguments

x
an ordination result

display
partial match to access scores for “sites” “species”, “loadings” or “xmatrix”. The latter two are only available for symcoca.

choices
numeric; the ordination axes to return.

scaling
logical; whether scores should be rescaled by the quarter root of the eigenvalues using rescale.symcoca.

...
arguments to be passed to other methods.

Details

Implements a scores method for symmetric co-correspondence analysis ordination results.

Value

A list with one or more components containing matrices of the requested scores:

species

A list with two components, Y and X, containing the species scores for the response matrix Y and the predictor matrix X respectively.

sites

A list with two components, Y and X, containing the site scores for the response matrix Y and the predictor matrix X respectively.

loadings

A list with two components, Y and X containing the loadings for the response and predictor matrix. For symcoca only.

xmatrix

The X matrix. For symcoca only.

References

ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence Analysis: a new ordination method to relate two community compositions. Ecology 85(3), 834--846

See also

Examples

od <- options(digits = 4)
## load some data
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model, a symmetric CoCA
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")

#> 
#> Removed some species that contained no data in: beetles, plants 

## extract the scores
scr <- scores(bp.sym)

## predictive CoCA using SIMPLS and formula interface
bp.pred <- coca(beetles ~ ., data = plants)

#> 
#> Removed some species that contained no data in: beetles, plants 

scr2 <- scores(bp.pred)

options(od)