Exctract CoCA model weights

Extractor function to identify and select appropriate analysis weights from Co-correspondence Analysis models.

# S3 method for class 'symcoca'
weights(object, ...)

# S3 method for class 'predcoca'
weights(object, ...)

Arguments

object

an object of class "symcoca" or "predcoca".

...

arguments passed to other methods. Not used.

Value

A numeric vector of common site weights is currently returned. These correspond to \(R_{0}\) in ter Braak and Schaffers (2004).

References

ter Braak, C.J.F and Schaffers, A.P. (2004) Co-Correspondence Analysis: a new ordination method to relate two community compositions. Ecology 85(3), 834–846.

Author

Gavin L. Simpson

See also

Examples

## symmetric CoCA
data(beetles)
data(plants)

## log transform the bettle data
beetles <- log(beetles + 1)

## fit the model
bp.sym <- coca(beetles ~ ., data = plants, method = "symmetric")
#> 
#> Removed some species that contained no data in: beetles, plants 

## weights == R[0]
weights(bp.sym)
#>  [1] 0.0146737 0.1586728 0.0905607 0.0052562 0.0052562 0.0182873 0.0373412
#>  [8] 0.0343846 0.1818879 0.0073368 0.0060228 0.0178493 0.0051467 0.0077749
#> [15] 0.0286903 0.0014236 0.0326325 0.0205869 0.0222295 0.0415024 0.0290188
#> [22] 0.0106220 0.0192729 0.0425975 0.0652650 0.0560666 0.0197109 0.0171923
#> [29] 0.0003285 0.0024091